Create a singleton Sound Manager and add sound to the game
Add Menu background music to be played when the player is in the Main
Menu, Options, or Help screens. Add Game music for when the player
enters the Game Activity, and if they finish the game it goes back to
the Menu music once they get back to the menu. During the Welcome
sequence, no music is playing, it starts once the user enters the Main
Menu (like I've seen in most games, it seems)

The logic behind the Sound Manager is that each activity has a
"keepPlaying" boolean variable that is false by default, but they set it
to true right before they go to another activity with the same music,
because then the onPause sees that keepPlaying is true and doesn't stop
the music. If the home button is pressed, however, then keepPlaying
was not set to true and the music will stop, and then it will resume
once you reenter the app in the onResume method. This way the music can
play continuously between certain activities with no pause/resume lag
and the playback structure/MediaPlayer object can be released when the
user leaves the app (as recommended by the Android docs)