Implement game saving functionality in Game Activity
Previously created the Save and Don't Save buttons when trying to leave
the Game Activity, but they don't do anything.
Make it so that clicking "Don't Save" clears the current saved game, and
the next time you press Play it will be a new game. When you click
"Save", it should save the game and exit to the main menu, and next time
you click play it will load the saved game. To discard a previously
saved game, you must open it and leave the game and click "Don't Save"